requirements.txt
streamlit>=1.28.0
pymongo>=4.0.0
pydantic>=2.0.0
pandas>=2.0.0
python-dotenv>=1.0.0
scikit-learn>=1.3.0
numpy>=1.24.0
kaggle>=1.5.0
seaborn>=0.12.0
wordcloud>=1.9.0
nltk>=3.8.0
regex>=2023.0.0
matplotlib>=3.7.0
joblib>=1.3.0
